+++ packages/e2e-tests/next-app-with-base-path/next.config.js
@@ -2,6 +2,9 @@ module.exports = {
   eslint: {
     ignoreDuringBuilds: true
   },
+  images: {
+    domains: ["raw.githubusercontent.com"]
+  },
   basePath: "/basepath",
   async redirects() {
     return [