+++ packages/libs/lambda-at-edge/src/build.ts
@@ -320,7 +320,8 @@ class Builder {
             const isNotStaticPropsJSONFile = path.extname(file) !== ".json";
             const isNotApiPage =
               separateApiLambda && !useV2Handler
-                ? pathToPosix(file).indexOf("pages/api") === -1
+                ? pathToPosix(file).indexOf("pages/api/") === -1 &&
+                  !pathToPosix(file).endsWith("pages/api") // don't copy api directory nor files in it, but copy pages that happen to start with pages/api
                 : true;
 
             // If there are API routes, include all JS files.
@@ -347,6 +348,7 @@ class Builder {
           }
         }
       ),
+      this.copyJSFiles(DEFAULT_LAMBDA_CODE_DIR),
       this.copyChunks(DEFAULT_LAMBDA_CODE_DIR),
       fse.copy(
         join(this.dotNextDir, "prerender-manifest.json"),
@@ -411,6 +413,7 @@ class Builder {
         join(this.serverlessDir, "pages/api"),
         join(this.outputDir, API_LAMBDA_CODE_DIR, "pages/api")
       ),
+      this.copyJSFiles(API_LAMBDA_CODE_DIR),
       this.copyChunks(API_LAMBDA_CODE_DIR),
       fse.writeJson(
         join(this.outputDir, API_LAMBDA_CODE_DIR, "manifest.json"),
@@ -438,6 +441,7 @@ class Builder {
         join(this.outputDir, REGENERATION_LAMBDA_CODE_DIR),
         !!this.buildOptions.minifyHandlers
       ),
+      this.copyJSFiles(REGENERATION_LAMBDA_CODE_DIR),
       this.copyChunks(REGENERATION_LAMBDA_CODE_DIR),
       fse.copy(
         join(this.serverlessDir, "pages"),
@@ -462,14 +466,34 @@ class Builder {
   /**
    * copy chunks if present and not using serverless trace
    */
-  copyChunks(buildDir: string): Promise<void> {
+  copyChunks(handlerDir: string): Promise<void> {
     return !this.buildOptions.useServerlessTraceTarget &&
       fse.existsSync(join(this.serverlessDir, "chunks"))
       ? fse.copy(
           join(this.serverlessDir, "chunks"),
-          join(this.outputDir, buildDir, "chunks")
+          join(this.outputDir, handlerDir, "chunks")
         )
       : Promise.resolve();
+  }
+
+  /**
+   * Copy additional JS files needed such as webpack-runtime.js (new in Next.js 12)
+   */
+  async copyJSFiles(handlerDir: string): Promise<void> {
+    await Promise.all([
+      (await fse.pathExists(join(this.serverlessDir, "webpack-api-runtime.js")))
+        ? fse.copy(
+            join(this.serverlessDir, "webpack-api-runtime.js"),
+            join(this.outputDir, handlerDir, "webpack-api-runtime.js")
+          )
+        : Promise.resolve(),
+      (await fse.pathExists(join(this.serverlessDir, "webpack-runtime.js")))
+        ? fse.copy(
+            join(this.serverlessDir, "webpack-runtime.js"),
+            join(this.outputDir, handlerDir, "webpack-runtime.js")
+          )
+        : Promise.resolve()
+    ]);
   }
 
   /**