+++ packages/serverless-components/aws-lambda/src/component.ts
@@ -11,6 +11,7 @@ import {
   configChanged,
   pack
 } from "./utils";
+import { waitUntilReady } from "./waitUntilReady";
 
 const outputsList = [
   "name",
@@ -97,6 +98,9 @@ class AwsLambda extends Component {
       const createResult = await createLambda({ lambda, ...config });
       config.arn = createResult.arn;
       config.hash = createResult.hash;
+
+      // Wait for Lambda to be in a ready state after creation and before doing anything else
+      await waitUntilReady(this.context, config.name, config.region);
     } else {
       config.arn = prevLambda.arn;
 
@@ -107,11 +111,17 @@ class AwsLambda extends Component {
           await updateLambdaCode({ lambda, ...config });
         }
 
+        // Wait for Lambda to be in a ready state after code updated and before doing anything else
+        await waitUntilReady(this.context, config.name, config.region);
+
         this.context.status(`Updating`);
         this.context.debug(`Updating ${config.name} lambda config.`);
 
         const updateResult = await updateLambdaConfig({ lambda, ...config });
         config.hash = updateResult.hash;
+
+        // Wait for Lambda to be ready after updating config and before doing anything else
+        await waitUntilReady(this.context, config.name, config.region);
       }
     }