+++ packages/libs/lambda-at-edge/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.5.3-alpha.5](https://github.com/serverless-nextjs/serverless-next.js/compare/v3.5.3-alpha.4...v3.5.3-alpha.5) (2021-11-08)
+
+**Note:** Version bump only for package @sls-next/lambda-at-edge
+
+## [3.5.3-alpha.4](https://github.com/serverless-nextjs/serverless-next.js/compare/v3.5.3-alpha.3...v3.5.3-alpha.4) (2021-10-30)
+
+### Bug Fixes
+
+- **lambda-at-edge:** properly include pages that start with pages/api (but are not api pages) ([#1994](https://github.com/serverless-nextjs/serverless-next.js/issues/1994)) ([26a756c](https://github.com/serverless-nextjs/serverless-next.js/commit/26a756cb413771bca7e649a1210b7a983ec569d0))
+
+## [3.5.3-alpha.0](https://github.com/serverless-nextjs/serverless-next.js/compare/v3.5.2...v3.5.3-alpha.0) (2021-10-29)
+
+**Note:** Version bump only for package @sls-next/lambda-at-edge
+
+## [3.5.2](https://github.com/serverless-nextjs/serverless-next.js/compare/v3.5.1...v3.5.2) (2021-10-28)
+
+**Note:** Version bump only for package @sls-next/lambda-at-edge
+
+## [3.5.1](https://github.com/serverless-nextjs/serverless-next.js/compare/v3.5.0...v3.5.1) (2021-10-28)
+
+**Note:** Version bump only for package @sls-next/lambda-at-edge
+
+# [3.5.0](https://github.com/serverless-nextjs/serverless-next.js/compare/v3.5.0-alpha.11...v3.5.0) (2021-10-28)
+
+**Note:** Version bump only for package @sls-next/lambda-at-edge
+
+# [3.5.0-alpha.11](https://github.com/serverless-nextjs/serverless-next.js/compare/v3.5.0-alpha.10...v3.5.0-alpha.11) (2021-10-27)
+
+### Features
+
+- **lambda-at-edge:** copy additional JS files to handlers ([#1970](https://github.com/serverless-nextjs/serverless-next.js/issues/1970)) ([e7c81ef](https://github.com/serverless-nextjs/serverless-next.js/commit/e7c81efde360bf7921c5aef088479154fa94280d))
+
 # [3.5.0-alpha.10](https://github.com/serverless-nextjs/serverless-next.js/compare/v3.5.0-alpha.9...v3.5.0-alpha.10) (2021-10-25)
 
 **Note:** Version bump only for package @sls-next/lambda-at-edge